+++ lib/qonfig.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require 'set'
 require 'yaml'
 require 'json'
 require 'erb'
@@ -63,4 +64,6 @@ module Qonfig
   register_plugin('toml', Qonfig::Plugins::TOML)
   # @since 0.19.0
   register_plugin('pretty_print', Qonfig::Plugins::PrettyPrint)
+  # @since 0.25.0
+  register_plugin('vault', Qonfig::Plugins::Vault)
 end