+++ spec/features/dot_notation_spec.rb
@@ -8,6 +8,8 @@ describe 'Dot-notation' do
           setting :cheburek, 'test'
         end
 
+        setting 'foo.bar', 100_500
+
         setting :frek do
           setting :jek do
             setting :bek, 123_456
@@ -21,18 +23,21 @@ describe 'Dot-notation' do
     expect(config.key?('kek.pek.cheburek')).to eq(true)
     expect(config.key?('kek.pek')).to eq(true)
     expect(config.key?('kek')).to eq(true)
+    expect(config.key?('kek.foo.bar')).to eq(true)
     expect(config.key?('kek.cheburek.pek')).to eq(false)
     expect(config.key?('kek.cheburek')).to eq(false)
 
     expect(config.option?('kek.pek.cheburek')).to eq(true)
     expect(config.option?('kek.pek')).to eq(true)
     expect(config.option?('kek')).to eq(true)
+    expect(config.option?('kek.foo.bar')).to eq(true)
     expect(config.option?('kek.cheburek.pek')).to eq(false)
     expect(config.option?('kek.cheburek')).to eq(false)
 
     expect(config.setting?('kek.pek.cheburek')).to eq(true)
     expect(config.setting?('kek.pek')).to eq(true)
     expect(config.setting?('kek')).to eq(true)
+    expect(config.setting?('kek.foo.bar')).to eq(true)
     expect(config.setting?('kek.cheburek.pek')).to eq(false)
     expect(config.setting?('kek.cheburek')).to eq(false)
   end
@@ -41,6 +46,7 @@ describe 'Dot-notation' do
     expect(config.dig('kek.pek.cheburek')).to eq('test')
     expect(config.dig('kek.pek')).to be_a(Qonfig::Settings)
     expect(config.dig('kek')).to be_a(Qonfig::Settings)
+    expect(config.dig('kek.foo.bar')).to eq(100_500)
 
     expect { config.dig('kek.pek.ululek') }.to raise_error(Qonfig::UnknownSettingError)
     expect { config.dig('kek.ululek') }.to raise_error(Qonfig::UnknownSettingError)
@@ -51,7 +57,8 @@ describe 'Dot-notation' do
     expect(config.subset('kek', 'kek.frek')).to match(
       'kek' => {
         'frek' => { 'jek' => { 'bek' => 123456 } },
-        'pek' => { 'cheburek'=>'test' }
+        'pek' => { 'cheburek'=>'test' },
+        'foo.bar' => 100_500
       },
       'frek' => {
         'jek' => { 'bek' => 123_456 }
@@ -59,13 +66,16 @@ describe 'Dot-notation' do
     )
 
     expect do
-      config.subset('kek', 'kek.frek', 'kek.lel')
+      config.subset('kek', 'kek.frek', 'kek.lel', 'kek.foo')
     end.to raise_error(Qonfig::UnknownSettingError)
   end
 
   specify '#slice' do
     expect(config.slice('kek.pek')).to match('pek' => { 'cheburek' => 'test' })
     expect(config.slice('kek.frek')).to match('frek' => { 'jek' => { 'bek' => 123_456 } })
+
+    # TODO: fix this
+    # expect(config.slice('kek.foo.bar')).to match('foo.bar' => 100_500)
 
     expect { config.slice('lel') }.to raise_error(Qonfig::UnknownSettingError)
     expect { config.slice('kek.lek') }.to raise_error(Qonfig::UnknownSettingError)
@@ -75,10 +85,12 @@ describe 'Dot-notation' do
     expect(config.slice_value('kek.pek.cheburek')).to eq('test')
     expect(config.slice_value('kek.pek')).to match('cheburek' => 'test')
     expect(config.slice_value('kek.frek')).to match('jek' => { 'bek' => 123_456 })
+    expect(config.slice_value('kek.foo.bar')).to eq(100_500)
 
     expect { config.slice_value('kek.pek.lelek') }.to raise_error(Qonfig::UnknownSettingError)
     expect { config.slice_value('kek.frek.bek') }.to raise_error(Qonfig::UnknownSettingError)
     expect { config.slice_value('lel') }.to raise_error(Qonfig::UnknownSettingError)
+    expect { config.slice_value('kek.foo') }.to raise_error(Qonfig::UnknownSettingError)
   end
 
   specify '#[]' do
@@ -86,9 +98,60 @@ describe 'Dot-notation' do
     expect(config['kek.pek']).to be_a(Qonfig::Settings)
     expect(config['kek.pek']['cheburek']).to eq('test')
     expect(config['kek']['frek.jek']['bek']).to eq(123_456)
+    expect(config['kek']['foo.bar']).to eq(100_500)
+    expect(config['kek.foo.bar']).to eq(100_500)
 
+    expect { config['kek.foo'] }.to raise_error(Qonfig::UnknownSettingError)
     expect { config['kek.pek.lelek'] }.to raise_error(Qonfig::UnknownSettingError)
     expect { config['kek.frek']['bek'] }.to raise_error(Qonfig::UnknownSettingError)
     expect { config['lel'] }.to raise_error(Qonfig::UnknownSettingError)
+  end
+
+  describe '#to_h' do
+    specify 'default behavior' do
+      expect(config.to_h(dot_style: true)).to match(
+        'kek.pek.cheburek' => 'test',
+        'kek.foo.bar' => 100_500,
+        'kek.frek.jek.bek' => 123_456
+      )
+    end
+
+    specify 'with key and value transformations' do
+      transformer = -> (value) { "#{value}!!" }
+
+      # transformations: key ONLY
+      hash = config.to_h(
+        dot_style: true,
+        key_transformer: transformer
+      )
+      expect(hash).to match(
+        'kek.pek.cheburek!!' => 'test',
+        'kek.foo.bar!!' => 100_500,
+        'kek.frek.jek.bek!!' => 123_456
+      )
+
+      # transformations: value ONLY
+      hash = config.to_h(
+        dot_style: true,
+        value_transformer: transformer
+      )
+      expect(hash).to match(
+        'kek.pek.cheburek' => 'test!!',
+        'kek.foo.bar' => '100500!!',
+        'kek.frek.jek.bek' => '123456!!'
+      )
+
+      # transformations: key AND value
+      hash = config.to_h(
+        dot_style: true,
+        key_transformer: transformer,
+        value_transformer: transformer
+      )
+      expect(hash).to match(
+        'kek.pek.cheburek!!' => 'test!!',
+        'kek.foo.bar!!' => '100500!!',
+        'kek.frek.jek.bek!!' => '123456!!'
+      )
+    end
   end
 end